+++ consensus/src/marshal/mocks/mod.rs
@@ -1,4 +1,3 @@
 pub mod application;
 pub mod block;
-pub mod epocher;
 pub mod resolver;
