+++ consensus/src/marshal/mod.rs
@@ -57,6 +57,7 @@
 
 pub mod actor;
 pub use actor::Actor;
+pub mod cache;
 pub mod config;
 pub use config::Config;
 pub mod finalizer;
